requirement.txt
numpy
pandas
opencv-python
tensorflow==2.11
pillow
flask